+++ tests/testthat/test-get_aux_table.R
@@ -1,6 +1,6 @@
 skip_if(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER") == "")
 lkups <- create_versioned_lkups(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER"))
-data_folder_root <-lkups$versions_paths[[lkups$latest_release]]$data_root
+data_folder_root <- lkups$versions_paths[[lkups$latest_release]]$data_root
 
 tables <- c("gdp", "pce", "pop", "cpi", "ppp")
 
@@ -9,4 +9,9 @@ test_that("get_aux_table() works", {
     get_aux_table(data_folder_root, table = x)
   })
   expect_equal(length(dl), length(tables))
+})
+
+test_that("get_aux_table() returns an error", {
+  expect_error(pipapi::get_aux_table(data_folder_root, "../survey_data/ARG_1980_EPH_D2_INC_GROUP.fst"),
+               "Error opening fst file for reading, please check access rights and file availability")
 })