+++ tests/testthat/test-pip.R
@@ -6,7 +6,7 @@ skip("Disable until a full set of anonymous package data has been created")
 
 # files <- sub("[.]fst", "", list.files("../testdata/app_data/20210401/survey_data/"))
 lkups <- create_versioned_lkups(Sys.getenv("PIPAPI_DATA_ROOT_FOLDER"))
-lkups <-lkups$versions_paths[[lkups$latest_release]]
+lkup <- lkups$versions_paths[[lkups$latest_release]]
 
 test_that("Reporting level filtering is working", {
   reporting_levels <- c("national", "urban", "rural", "all")
@@ -20,7 +20,7 @@ test_that("Reporting level filtering is working", {
                       reporting_level = x,
                       fill_gaps = FALSE,
                       ppp = 10,
-                      lkup = lkups,
+                      lkup = lkup,
                       debug = FALSE)
                 })
   names(tmp) <- reporting_levels
@@ -39,8 +39,8 @@ test_that("Reporting level filtering is working", {
   })
 
 # Use only test data
-# lkups$svy_lkup <- lkups$svy_lkup[(cache_id %in% files | country_code == "AGO")]
-# lkups$ref_lkup <- lkups$ref_lkup[(cache_id %in% files | country_code == "AGO")]
+# lkup$svy_lkup <- lkup$svy_lkup[(cache_id %in% files | country_code == "AGO")]
+# lkup$ref_lkup <- lkup$ref_lkup[(cache_id %in% files | country_code == "AGO")]
 
 # Check output type ----
 test_that("output type is correct", {
@@ -48,7 +48,7 @@ test_that("output type is correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups
+    lkup = lkup
   )
 
   expect_equal(class(tmp), c("data.table", "data.frame"))
@@ -56,17 +56,17 @@ test_that("output type is correct", {
 
 # Check empty response
 test_that("empty response is returned if no metadata is found", {
-  tmp <- pip("COL", year = 2050, lkup = lkups)
+  tmp <- pip("COL", year = 2050, lkup = lkup)
   expect_equal(nrow(tmp), 0)
-  tmp <- pip("COL", year = 2050, lkup = lkups, fill_gaps = TRUE)
+  tmp <- pip("COL", year = 2050, lkup = lkup, fill_gaps = TRUE)
   expect_equal(nrow(tmp), 0)
 })
 
 # Check response columns
 test_that("returned columns are the same for all non-group_by queries", {
-  tmp1 <- pip('AGO', 2000, lkup = lkups)
-  tmp2 <- pip('AGO', 2010, lkup = lkups, fill_gaps = TRUE)
-  tmp3 <- pip('AGO', 2050, lkup = lkups)
+  tmp1 <- pip('AGO', 2000, lkup = lkup)
+  tmp2 <- pip('AGO', 2010, lkup = lkup, fill_gaps = TRUE)
+  tmp3 <- pip('AGO', 2050, lkup = lkup)
   expect_identical(names(tmp1), names(tmp2))
   expect_identical(names(tmp1), names(tmp3))
   # skip("collapsed columns (e.g. survey_year, cpi) are converted to character")
@@ -84,9 +84,9 @@ test_that("year selection is working", {
     country = "AGO",
     year = "all",
     povline = 1.9,
-    lkup = lkups
+    lkup = lkup
   )
-  check <- sum(lkups$svy_lkup$country_code == "AGO")
+  check <- sum(lkup$svy_lkup$country_code == "AGO")
   expect_equal(nrow(tmp), check)
 
   # Most recent year for a single country
@@ -94,9 +94,9 @@ test_that("year selection is working", {
     country = "AGO",
     year = "mrv",
     povline = 1.9,
-    lkup = lkups
+    lkup = lkup
   )
-  check <- max(lkups$svy_lkup[country_code == "AGO"]$reporting_year)
+  check <- max(lkup$svy_lkup[country_code == "AGO"]$reporting_year)
   expect_equal(tmp$reporting_year, sum(check))
 
   # Most recent year for a single country (w/ fill_gaps)
@@ -105,9 +105,9 @@ test_that("year selection is working", {
     year = "mrv",
     povline = 1.9,
     fill_gaps = TRUE,
-    lkup = lkups
+    lkup = lkup
   )
-  check <- max(lkups$ref_lkup$reporting_year)
+  check <- max(lkup$ref_lkup$reporting_year)
   expect_equal(tmp$reporting_year, check)
 
   # Most recent year for all countries
@@ -115,9 +115,9 @@ test_that("year selection is working", {
     country = "all",
     year = "mrv",
     povline = 1.9,
-    lkup = lkups
+    lkup = lkup
   )
-  check <- max(lkups$svy_lkup$reporting_year)
+  check <- max(lkup$svy_lkup$reporting_year)
   expect_equal(unique(tmp$reporting_year), check)
 
 })
@@ -128,7 +128,7 @@ test_that("welfare_type selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     welfare_type = "all"
   )
 
@@ -138,7 +138,7 @@ test_that("welfare_type selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     welfare_type = "consumption"
   )
 
@@ -148,7 +148,7 @@ test_that("welfare_type selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     welfare_type = "income"
   )
 
@@ -161,7 +161,7 @@ test_that("reporting_level selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     reporting_level = "all"
   )
 
@@ -171,7 +171,7 @@ test_that("reporting_level selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     reporting_level = "national"
   )
 
@@ -181,7 +181,7 @@ test_that("reporting_level selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     reporting_level = "rural"
   )
 
@@ -191,7 +191,7 @@ test_that("reporting_level selection are correct", {
     country = "all",
     year = "all",
     povline = 3.5,
-    lkup = lkups,
+    lkup = lkup,
     reporting_level = "urban"
   )
 
@@ -206,7 +206,7 @@ test_that("Aggregation is working", {
     year = "all",
     povline = 3.5,
     aggregate = TRUE,
-    lkup = lkups
+    lkup = lkup
   )
   expect_equal(nrow(tmp), 1)
 })
@@ -218,7 +218,7 @@ test_that("Imputation is working", {
     year = "all",
     povline = 3.5,
     fill_gaps = TRUE,
-    lkup = lkups
+    lkup = lkup
   )
   # Why is this correct? E.g. tmp %>% group_by(country_code) %>% summarise(n = n())
   expect_equal(nrow(tmp), 195)
@@ -231,7 +231,7 @@ test_that("Imputation is working for mixed distributions aggregate / micro", {
     year = 1993,
     povline = 1.9,
     fill_gaps = TRUE,
-    lkup = lkups
+    lkup = lkup
   )
 
   expect_equal(nrow(tmp), 3)
@@ -247,7 +247,7 @@ test_that("Imputation is working for mixed distributions group / micro", {
     year = 2015,
     povline = 1.9,
     fill_gaps = TRUE,
-    lkup = lkups
+    lkup = lkup
   )
 
   expect_equal(nrow(tmp), 1)
@@ -262,7 +262,7 @@ test_that("imputation is working for extrapolated aggregate distribution", {
     year = 1988,
     povline = 1.9,
     fill_gaps = TRUE,
-    lkup = lkups
+    lkup = lkup
   )
 
   expect_equal(nrow(tmp), 3)
@@ -275,29 +275,29 @@ test_that("imputation is working for extrapolated aggregate distribution", {
 test_that("Distributional stats are correct for interpolated/extrapolated reporting years",{
 
   # Extrapolation (one year)
-  tmp1 <- pip("AGO", year = 1981, fill_gaps = TRUE, lkup = lkups)
-  tmp2 <- pip("AGO", year = 2000, fill_gaps = FALSE, lkup = lkups)
+  tmp1 <- pip("AGO", year = 1981, fill_gaps = TRUE, lkup = lkup)
+  tmp2 <- pip("AGO", year = 2000, fill_gaps = FALSE, lkup = lkup)
   expect_equal(tmp1$gini, tmp2$gini)
   expect_equal(tmp1$median, tmp2$median)
   expect_equal(tmp1$mld, tmp2$mld)
   expect_equal(tmp1$decile10, tmp2$decile10)
 
   # Interpolation (one year)
-  tmp1 <- pip("AGO", year = 2004, fill_gaps = TRUE, lkup = lkups)
+  tmp1 <- pip("AGO", year = 2004, fill_gaps = TRUE, lkup = lkup)
   expect_equal(tmp1$gini, NA_real_)
   expect_equal(tmp1$median ,NA_real_)
   expect_equal(tmp1$mld, NA_real_)
   expect_equal(tmp1$decile10, NA_real_)
 
   # Extrapolation (multiple years)
-  tmp1 <- pip("AGO", year = 1981:1999, fill_gaps = TRUE, lkup = lkups)
+  tmp1 <- pip("AGO", year = 1981:1999, fill_gaps = TRUE, lkup = lkup)
   expect_equal(unique(tmp1$gini), tmp2$gini)
   expect_equal(unique(tmp1$median), tmp2$median)
   expect_equal(unique(tmp1$mld), tmp2$mld)
   expect_equal(unique(tmp1$decile10), tmp2$decile10)
 
   # Interpolation (mulitiple year)
-  tmp1 <- pip("AGO", year = 2001:2007, fill_gaps = TRUE, lkup = lkups)
+  tmp1 <- pip("AGO", year = 2001:2007, fill_gaps = TRUE, lkup = lkup)
   expect_equal(unique(tmp1$gini), NA_real_)
   expect_equal(unique(tmp1$median), NA_real_)
   expect_equal(unique(tmp1$mld), NA_real_)
@@ -313,7 +313,7 @@ test_that("Regional aggregations are working", {
     year = "2000",
     group_by = "wb",
     povline = 3.5,
-    lkup = lkups
+    lkup = lkup
   )
 
   expect_equal(nrow(tmp), 3) # Should be changed if lkups are updated. Full set of regions is 8.
@@ -325,9 +325,36 @@ test_that("pop_share option is working", {
     country = "AGO",
     year = 2000,
     popshare = .2,
-    lkup = lkups
+    lkup = lkup
   )
 
   expect_equal(nrow(tmp), 1)
 })
 
+#Check pip country name case insensitive
+
+test_that("pip country name case insensitive", {
+  #Run it on pip-fake-data
+  tmp1 <- pip(country = "nga",year = "ALL", povline = 1.9, lkup = lkup)
+  tmp2 <- pip(country = "NGA",year = "all", povline = 1.9, lkup = lkup)
+  tmp3 <- pip(country = "All",year = "ALL", povline = 1.9, lkup = lkup)
+  tmp4 <- pip(country = "chn",year = "1981", povline = 1.9, lkup = lkup)
+  tmp5 <- pip(country = "chn",year = "ALL", povline = 1.9, lkup = lkup)
+
+  expect_equal(nrow(tmp1), 1)
+  expect_equal(nrow(tmp2), 1)
+  expect_equal(nrow(tmp3), 22)
+  expect_equal(nrow(tmp4), 3)
+  expect_equal(nrow(tmp5), 6)
+})
+
+
+#Better error message when more than one data set is passed.
+
+test_that("error when more than one dataset is passed", {
+
+  expect_error(pip(country = "all", year = "all", povline = 1.9, lkup = lkups),
+               "You are probably passing more than one dataset as lkup argument.
+  Try passing a single one by subsetting it lkup <- lkups$versions_paths$dataset_name_PROD",
+  fixed = TRUE)
+})